Added common library
