+++ atcoder-problems-scraper/src/lib.rs
@@ -1,9 +1,5 @@
-#[macro_use]
-extern crate diesel;
-
-pub mod models;
-pub mod schema;
 pub mod scraper;
 pub mod sql;
 
-pub use models::{Contest, Problem, Submission};
+pub use atcoder_problems_sql_common::models::{Contest, Problem, Submission};
+pub use atcoder_problems_sql_common::schema;
