+++ atcoder-problems-sql-common/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "atcoder-problems-sql-common"
+version = "0.1.0"
+authors = ["kenkoooo <kenkou.n@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+diesel = { version = "1.4.2", features = ["postgres"] }
