+++ atcoder-problems-sql-common/src/lib.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+extern crate diesel;
+
+pub mod schema;
+pub mod models;
