+++ atcoder-problems-sql-common/src/models.rs
@@ -1,0 +1,31 @@
+use super::schema::{contests, problems, submissions};
+
+#[derive(Debug, Eq, PartialEq, Queryable, Insertable)]
+pub struct Contest {
+    pub id: String,
+    pub start_epoch_second: i64,
+    pub duration_second: i64,
+    pub title: String,
+    pub rate_change: String,
+}
+
+#[derive(Debug, Eq, PartialEq, Queryable, Insertable)]
+pub struct Problem {
+    pub id: String,
+    pub contest_id: String,
+    pub title: String,
+}
+
+#[derive(Debug, Queryable, Insertable, Clone)]
+pub struct Submission {
+    pub id: i64,
+    pub epoch_second: i64,
+    pub problem_id: String,
+    pub contest_id: String,
+    pub user_id: String,
+    pub language: String,
+    pub point: f64,
+    pub length: i32,
+    pub result: String,
+    pub execution_time: Option<i32>,
+}
