+++ atcoder-problems-sql-common/src/schema.rs
